pyproject.toml
[tool.poetry]
name = "cfb-betting-model"
version = "0.1.0"
description = "Production-ready CFB betting model for ATS, moneyline, and totals predictions"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.0"
numpy = "^2.0.0"
pyarrow = "^18.0.0"
pydantic = "^2.0.0"
requests = "^2.31.0"
tenacity = "^8.2.0"
python-dotenv = "^1.0.0"
typer = {extras = ["all"], version = "^0.9.0"}
xgboost = "^2.0.0"
scikit-learn = "^1.3.0"
meteostat = "^1.6.0"
geopy = "^2.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.0"
black = "^23.0.0"
isort = "^5.12.0"
pytest = "^7.4.0"
pre-commit = "^3.4.0"

[tool.poetry.scripts]
cfb-model = "src.cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

